requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
tqdm>=4.66.0
python-dotenv>=1.0.0
flask>=3.0.0
anthropic>=0.25.0
psycopg2-binary>=2.9.0
